+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "expressive-mvc",
+	"name": "expressive-state",
 	"license": "MIT",
 	"private": true,
 	"workspaces": [
