+++ packages/preact/package.json
@@ -8,7 +8,7 @@
     "url": "https://github.com/gabeklein"
   },
   "license": "MIT",
-  "repository": "https://github.com/gabeklein/expressive-mvc",
+  "repository": "https://github.com/gabeklein/expressive-state",
   "publishConfig": {
     "access": "public"
   },
@@ -20,7 +20,7 @@
     "preversion": "npm run test && npm run build"
   },
   "dependencies": {
-    "@expressive/mvc": "^0.71.0",
+    "@expressive/state": "^0.71.0",
     "@expressive/react": "^0.71.0"
   },
   "devDependencies": {
