+++ packages/react/src/index.ts
@@ -1,4 +1,4 @@
-import { State } from '@expressive/mvc';
+import { State } from '@expressive/state';
 
 import { createElement, useEffect, useState } from 'react';
 
@@ -12,9 +12,9 @@ Pragma.createElement = createElement;
  * Augmented State namespace to include ReactState hooks as static methods.
  *
  * This retroactively adds React support to agnostic State defined
- * by component libraries which import `@expressive/mvc` directly.
+ * by component libraries which import `@expressive/state` directly.
  */
-declare module '@expressive/mvc' {
+declare module '@expressive/state' {
   namespace State {
     export import as = ReactState.as;
     export import get = ReactState.get;
@@ -27,7 +27,7 @@ State.use = ReactState.use;
 State.as = ReactState.as;
 
 export { ReactState as State, ReactState as default };
-export { Context, Observable, get, use, ref, set } from '@expressive/mvc';
+export { Context, Observable, get, use, ref, set } from '@expressive/state';
 export { Consumer, Provider } from './context';
 export { Fragment, createElement } from 'react';
 export { Pragma };
