+++ packages/solid/package.json
@@ -8,7 +8,7 @@
     "url": "https://github.com/gabeklein"
   },
   "license": "MIT",
-  "repository": "https://github.com/gabeklein/expressive-mvc",
+  "repository": "https://github.com/gabeklein/expressive-state",
   "publishConfig": {
     "access": "public"
   },
@@ -16,7 +16,7 @@
   "module": "src",
   "types": "src",
   "dependencies": {
-    "@expressive/mvc": "^0.71.0"
+    "@expressive/state": "^0.71.0"
   },
   "devDependencies": {
     "@solidjs/testing-library": "^0.8.8",
