+++ packages/solid/src/index.ts
@@ -1,13 +1,13 @@
-import { State } from '@expressive/mvc';
+import { State } from '@expressive/state';
 import { SolidState } from './state';
 
 /**
  * Augmented State namespace to include SolidState hooks as static methods.
  *
  * This retroactively adds Solid support to agnostic State defined
- * by component libraries which import `@expressive/mvc` directly.
+ * by component libraries which import `@expressive/state` directly.
  */
-declare module '@expressive/mvc' {
+declare module '@expressive/state' {
   namespace State {
     export import get = SolidState.get;
     export import use = SolidState.use;
@@ -19,4 +19,4 @@ State.use = SolidState.use;
 
 export { Provider } from './context';
 export { SolidState as State, SolidState as default };
-export { Context, get, use, ref, set } from '@expressive/mvc';
+export { Context, get, use, ref, set } from '@expressive/state';
